pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdrive-toolkit"
version = "0.1.0"
description = "A lightweight Google Drive toolkit for Kaggle, Colab, and local environments"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["google-drive", "pydrive2", "kaggle", "colab", "automation"]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydrive2>=1.15.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/gdrive-toolkit"
Issues = "https://github.com/yourusername/gdrive-toolkit/issues"
